// src/api/types.d.ts
export type ApiViewMasterServiceParam = {
  page: number,
  size: number
}

export type RawResultSearchMasterService = ViewedMasterService

export interface ViewedMasterService {
  updatedAt:          Date;
  createdAt:          Date;
  id:                 number;
  appliedNation:      AppliedNation[];
  appliedCompanyType: AppliedCompanyType[];
  serviceType:        string;
  serviceName:        string;
  serviceDescription: string;
  enable:             number;
  serviceStep:        ServiceStep[];
  serviceCycle:       ServiceCycle[];
}

export interface AppliedCompanyType {
  id:          number;
  companyType: string;
}

export interface AppliedNation {
  id:     number;
  nation: string;
}

export interface ServiceCycle {
  id:            number;
  cycleNumber:   number;
  pricePerCycle: number;
}

export interface ServiceStep {
  id:                      number;
  stepNo:                  number;
  name:                    string;
  estimatedCompletionTime: string;
  description:             string;
  documentRequired:        DocumentRequired[];
  result:                  Result[];
}

export interface DocumentRequired {
  id:               number;
  documentRequired: string;
}

export interface Result {
  id:     number;
  result: string;
}
// ...
export type ApiSearchMasterServiceParam = {
  "serviceId": string,
  "serviceName": string,
  "status": string,
  "appliedNation": string
}